Fixes & cleanup
