+++ app/src/main/kotlin/com/flxrs/dankchat/data/NotificationService.kt
@@ -306,7 +306,7 @@ class NotificationService : Service(), CoroutineScope {
     ) {
         companion object {
             fun Message.toNotificationData(): NotificationData? = when (this) {
-                is PrivMessage    -> if (highlightState?.isMention == true) NotificationData(channel, name, originalMessage) else null
+                is PrivMessage    -> if (highlights.hasMention()) NotificationData(channel, name, originalMessage) else null
                 is WhisperMessage -> NotificationData(
                     channel = "",
                     name = name,
