+++ app/src/main/kotlin/com/flxrs/dankchat/data/twitch/message/HighlightState.kt
@@ -1,12 +1,13 @@
 package com.flxrs.dankchat.data.twitch.message
 
-data class HighlightState(
+data class Highlight(
     val type: HighlightType,
     val customColor: Int? = null
 ) {
-    // TODO Check if enough
     val isMention = type == HighlightType.Username || type == HighlightType.Custom
 }
+
+fun List<Highlight>.hasMention(): Boolean = any(Highlight::isMention)
 
 enum class HighlightType {
     Username,
