+++ app/src/main/kotlin/com/flxrs/dankchat/data/twitch/message/SystemMessage.kt
@@ -1,0 +1,10 @@
+package com.flxrs.dankchat.data.twitch.message
+
+import java.util.UUID
+
+data class SystemMessage(
+    val type: SystemMessageType,
+    override val timestamp: Long = System.currentTimeMillis(),
+    override val id: String = UUID.randomUUID().toString(),
+    override val highlights: List<Highlight> = emptyList(),
+) : Message()
